consensus_encoding: Move errors to submodule
In order to make the HTML docs render more clearly it would be nice to
have all the errors separate. We recently introduced a policy for
doing this repo wide.

Add a public `error` submodule and re-export all errors at the crate
root.

Because of anomalies in how `cargo` renders docs for re-exports
between types and modules explicitly use `doc(inline)` and `no_inline`
so we get exactly the behaviour we desire.
